+++ Assets/LJH/LJH_scripts/UI/ResultBarChartView.cs
@@ -13,14 +13,14 @@ namespace InterviewDb.UI
     public class ResultBarChartView : MonoBehaviour
     {
         [Header("5점 척도 저장 데이터 (음성 / 내용 / 태도)")]
-        [SerializeField] private double? audioScore;
-        [SerializeField] private double? contentScore;
-        [SerializeField] private double? attitudeScore;
+        [SerializeField] private int? audioScore;      // double? -> int?
+        [SerializeField] private int? contentScore;    // double? -> int?
+        [SerializeField] private int? attitudeScore;   // double? -> int?
 
         // 외부(UI 렌더러 등)에서 읽을 수 있는 읽기 전용 프로퍼티
-        public double? AudioScore => audioScore;
-        public double? ContentScore => contentScore;
-        public double? AttitudeScore => attitudeScore;
+        public int? AudioScore => audioScore;          // double? -> int?
+        public int? ContentScore => contentScore;      // double? -> int?
+        public int? AttitudeScore => attitudeScore;    // double? -> int?
 
         /// <summary>
         /// DB 세션 리포트(SessionReportRow) 객체에서 3대 점수 데이터를 추출하여 저장.
@@ -39,7 +39,7 @@ namespace InterviewDb.UI
         /// <summary>
         /// 영역별 점수 데이터를 직접 저장.
         /// </summary>
-        public void StoreScores(double? audio, double? content, double? attitude)
+        public void StoreScores(int? audio, int? content, int? attitude) // ★ 매개변수 int? 변경
         {
             audioScore = audio;
             contentScore = content;
@@ -60,7 +60,7 @@ namespace InterviewDb.UI
         // 기존 타 스크립트 호출 호환성 유지용 (Alias)
         // ------------------------------------------------------------
         public void DisplayReport(SessionReportRow report) => StoreReport(report);
-        public void SetScores(double? audioScore, double? contentScore, double? attitudeScore)
+        public void SetScores(int? audioScore, int? contentScore, int? attitudeScore) // 매개변수 int? 변경
             => StoreScores(audioScore, contentScore, attitudeScore);
     }
 }
